requirements.txt
numpy
pyelftools==0.27
tqdm
pyserial
